pyproject.toml
[project]
name = "toolpickr"
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu",
    "numpy",
]
[project.optional-dependencies]
gemini = ["google-genai"]